Move terrain generation out of WorldSim & use TerrainChunk for all terrain functionality
