+++ src/game/state.rs
@@ -11,6 +11,7 @@ use amethyst::{
 };
 
 use libdwarf::{actions::Action, resources::TaskQueue, world::WorldSim, Point3};
+use libterrain::TerrainGenerator;
 
 use crate::game::{
     components::{CameraFollow, Cursor, CursorSelected, Object, Player},
@@ -45,7 +46,8 @@ impl SimpleState for RunningState {
         };
 
         // Initialize simulation
-        WorldSim::new(world, map_width, map_height);
+        let terrain_gen = TerrainGenerator::new(map_width, map_height).build();
+        WorldSim::new(world, &terrain_gen.get_terrain(), map_width, map_height);
         // Render map
         let map_render = MapRenderer::initialize(world);
         initialize_camera(world, map_render.place(8, 8, 42, 0.0), self.zoom);
